Move GlobalAllocLock to gchelpers.cpp and switch to preemptive mode when spinning
